pyproject.toml
[project]
name = "garopabus"
version = "0.1.0"
description = "Sistema de informações de horários de ônibus"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django==5.1.3",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "django-cors-headers==4.6.0",
    "django-environ==0.11.2",
    "django-filter==24.3",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "pywebpush==2.0.3",
    "drf-yasg==1.21.8",
    "gunicorn==23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
